pyproject.toml
[tool.poetry]
name = "hercule"
version = "0.0.1"
description = "Put your browser under steroïds"
authors = ["Henri Chabert <chabhenrib@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
gunicorn = "^23.0.0"
pydantic-settings = "^2.6.1"
pytest = "^8.3.3"
faker = "^33.0.0"
httpx = "^0.27.2"
aiosqlite = "^0.20.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"
greenlet = "^3.1.1"
sqlmodel = "^0.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# pyproject.toml
[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
